Divided code into modules responsible for separately scraping recipes and loading/saving data.
